Added file creating in DocumentWriter and BinaryDocumentWriter
Took 4 minutes

+++ prematiclibraries-document/src/main/java/net/prematic/libraries/document/io/DocumentWriter.java
@@ -46,8 +46,11 @@ public interface DocumentWriter {
 
     default void write(File location, Document document, boolean pretty){
         try {
+            if(!location.exists()){
+                if(!location.createNewFile()) throw new IllegalArgumentException("Could not create "+location.getName());
+            }
             write(new FileOutputStream(location),document,pretty);
-        } catch (FileNotFoundException exception) {
+        } catch (IOException exception) {
             throw new IORuntimeException(exception);
         }
     }
